pyproject.toml
[tool.poetry]
name = "ray-worker"
version = "0.1.0"
description = ""
authors = ["Ken <ken@blackcape.io>"]

[tool.poetry.dependencies]
python = "3.8.13"
ray = {extras = ["core"], version = "^1.12.0"}
kafka-python = "<2.0.0"
pydantic = "^1.9.0"
fastapi = "^0.77.1"
toml = "^0.10.2"
minio = "^7.1.8"
asyncpg = "^0.25.0"
asyncpg-utils = "^0.6.0"
uvicorn = "^0.17.6"
psutil = "^5.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
